refactor: 메인페이지 카드 상세페이지와 경로 연결 및 뒤로가기 버튼 공통컴포넌트 분리
