--- src/types/home.ts
@@ -1,28 +1,0 @@
-export interface DuPickItem {
-  id: string;
-  name: string;
-  date: string;
-  location: string;
-}
-
-export interface ExhibitionCardData {
-  id: string;
-  title: string;
-  school: string;
-  period: string;
-}
-
-export interface ArtworkPreviewItem {
-  id: string;
-  name: string;
-  date: string;
-}
-
-export interface LoungePost {
-  id: string;
-  tag: string;
-  content: string;
-  author: string;
-  time: string;
-  views: string;
-}
